refactor(suite): Rename suite-common/analytics-types to suite-common/analytics
