+++ packages/suite/tsconfig.json
@@ -13,10 +13,10 @@
     "include": [".", "e2e", "**/*.json"],
     "references": [
         {
-            "path": "../../suite-common/analytics-redux"
+            "path": "../../suite-common/analytics-types"
         },
         {
-            "path": "../../suite-common/analytics-types"
+            "path": "../../suite-common/analytics-redux"
         },
         { "path": "../../suite-common/assets" },
         {
