+++ suite-common/analytics-types/package.json
@@ -1,5 +1,5 @@
 {
-    "name": "@suite-common/analytics-types",
+    "name": "@suite-common/analytics",
     "version": "1.0.0",
     "private": true,
     "license": "See LICENSE.md in repo root",
