+++ suite-common/connect-popup/package.json
@@ -11,7 +11,7 @@
     },
     "dependencies": {
         "@reduxjs/toolkit": "2.11.0",
-        "@suite-common/analytics-types": "workspace:*",
+        "@suite-common/analytics": "workspace:*",
         "@suite-common/redux-utils": "workspace:*",
         "@suite-common/toast-notifications": "workspace:*",
         "@suite-common/wallet-config": "workspace:*",
