+++ suite-common/redux-utils/package.json
@@ -12,7 +12,7 @@
     },
     "dependencies": {
         "@reduxjs/toolkit": "2.11.0",
-        "@suite-common/analytics-types": "workspace:*",
+        "@suite-common/analytics": "workspace:*",
         "@suite-common/metadata-types": "workspace:*",
         "@suite-common/platform-encryption": "workspace:*",
         "@suite-common/suite-rbf-labels-migrations-types": "workspace:*",
