+++ suite-common/walletconnect/src/walletConnectThunks.ts
@@ -8,7 +8,7 @@ import {
     populateAuthPayload,
 } from '@walletconnect/utils';
 
-import { EventType } from '@suite-common/analytics-types';
+import { EventType } from '@suite-common/analytics';
 import * as trezorConnectPopupActions from '@suite-common/connect-popup';
 import { createThunk } from '@suite-common/redux-utils';
 import { notificationsActions } from '@suite-common/toast-notifications';
