+++ suite-native/analytics/src/events/earnStakeTilePressedEvent.ts
@@ -1,4 +1,4 @@
-import type { EventDef } from '@suite-common/analytics-types';
+import type { EventDef } from '@suite-common/analytics';
 
 import { EventType } from '../constants';
