+++ suite-native/analytics/src/events/settingsChangeCoinBackendEvent.ts
@@ -1,4 +1,4 @@
-import type { AttributeDef, EventDef } from '@suite-common/analytics-types';
+import type { AttributeDef, EventDef } from '@suite-common/analytics';
 import { BackendType, NetworkSymbol } from '@suite-common/wallet-config';
 
 import { EventType } from '../constants';
