+++ suite-native/bluetooth/src/hooks/useBluetoothAdapter.ts
@@ -2,7 +2,7 @@ import { useEffect } from 'react';
 import { AppState } from 'react-native';
 import { useDispatch, useSelector } from 'react-redux';
 
-import { EventType } from '@suite-common/analytics-types';
+import { EventType } from '@suite-common/analytics';
 import { bluetoothActions, parseManufacturerData } from '@suite-common/bluetooth';
 import { useTranslate } from '@suite-native/intl';
 import { useAnalytics } from '@suite-native/services';
