+++ suite-native/device/package.json
@@ -14,7 +14,7 @@
         "@mobily/ts-belt": "^3.13.1",
         "@react-navigation/native": "7.1.26",
         "@reduxjs/toolkit": "2.11.0",
-        "@suite-common/analytics-types": "workspace:*",
+        "@suite-common/analytics": "workspace:*",
         "@suite-common/bluetooth": "workspace:*",
         "@suite-common/device-authenticity": "workspace:*",
         "@suite-common/firmware-authenticity": "workspace:*",
