+++ suite-native/module-authorize-device/package.json
@@ -13,7 +13,7 @@
         "@react-navigation/native": "7.1.26",
         "@react-navigation/native-stack": "7.9.0",
         "@reduxjs/toolkit": "2.11.0",
-        "@suite-common/analytics-types": "workspace:*",
+        "@suite-common/analytics": "workspace:*",
         "@suite-common/bluetooth": "workspace:*",
         "@suite-common/redux-utils": "workspace:*",
         "@suite-common/thp": "workspace:*",
