+++ suite-native/module-device-settings/src/hooks/useChangeDeviceName.ts
@@ -2,7 +2,7 @@ import { useSelector } from 'react-redux';
 
 import { useNavigation } from '@react-navigation/native';
 
-import { EventType } from '@suite-common/analytics-types';
+import { EventType } from '@suite-common/analytics';
 import { selectSelectedDevice } from '@suite-common/wallet-core';
 import { useForm } from '@suite-native/forms';
 import { useTranslate } from '@suite-native/intl';
