+++ suite-native/module-onboarding/package.json
@@ -14,7 +14,7 @@
         "@react-navigation/native": "7.1.26",
         "@react-navigation/native-stack": "7.9.0",
         "@reduxjs/toolkit": "2.11.0",
-        "@suite-common/analytics-types": "workspace:*",
+        "@suite-common/analytics": "workspace:*",
         "@suite-native/analytics": "workspace:*",
         "@suite-native/app-init": "workspace:*",
         "@suite-native/atoms": "workspace:*",
