+++ suite-native/thp/package.json
@@ -12,7 +12,7 @@
     "dependencies": {
         "@react-navigation/native": "7.1.26",
         "@reduxjs/toolkit": "2.11.0",
-        "@suite-common/analytics-types": "workspace:*",
+        "@suite-common/analytics": "workspace:*",
         "@suite-common/redux-utils": "workspace:*",
         "@suite-common/thp": "workspace:*",
         "@suite-common/wallet-core": "workspace:*",
