+++ suite/analytics/README.md
@@ -2,4 +2,4 @@
 
 This package is intended to be used only by Suite on the desktop (web) environment.
 
-All desktop-related events are defined here. For help how to define events or use them please refer to [`@suite-common/analytics-types`](../../suite-common/analytics-types/README.md).
+All desktop-related events are defined here. For help how to define events or use them please refer to [`@suite-common/analytics`](../../suite-common/analytics-types/README.md).
